implmenting livewire

+++ routes/web.php
@@ -13,12 +13,6 @@ use App\Http\Controllers\SubmitReportController;
 use App\Http\Controllers\RenterController;
 use Illuminate\Support\Facades\DB;
 
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
-use App\Models\User;
-use App\Models\Category;
-use App\Models\Images;
-use App\Models\Report;
 
 use Illuminate\Support\Facades\Route;
 
@@ -35,17 +29,17 @@ use Illuminate\Support\Facades\Route;
 // specify landing page 
 Route::middleware(['auth','verified'])->group( function(){
     Route::group(['middleware' => ['isAdmin']], function () {
-        Route::get('/admin/dashboard',[AdminDashboardController::class, 'index'])->name('dashboard');
+        Route::get('/dashboard',[AdminDashboardController::class, 'index'])->name('dashboard');
         Route::get('/admin/users', [UserController::class, 'index'])->name('admin.users.index');
 
         // controller group for category
         Route::controller(CategoryController::class)->group(function() {
             Route::get('/admin/categories/index', 'index')->name('admin.categories.index');
-            Route::post('/admin/categories/index', 'create')->name('categories.create');
+            Route::post('/admin/categories/index', 'create')->name('admin.categories.create');
             Route::get('/admin/categories/{categories}/edit', 'edit')->name('admin.categories.edit');
-            Route::put('/admin/categories/{categories}/update', 'update')->name('categories.update');
-            Route::delete('/admin/categories/{categories}/delete', 'delete')->name('categories.delete');
-            Route::get('/admin/categories/restore','restore')->name('categories.restore');
+            Route::put('/admin/categories/{categories}/update', 'update')->name('admin.categories.update');
+            Route::delete('/admin/categories/{categories}/delete', 'delete')->name('admin.categories.delete');
+            Route::get('/admin/categories/restore','restore')->name('admin.categories.restore');
         });
         // controller group for report
         Route::controller(ReportController::class)->group(function() {
@@ -53,11 +47,11 @@ Route::middleware(['auth','verified'])->group( function(){
         });
         // controller group for category
         Route::controller(AppartmentController::class)->group(function() {
-            Route::get('/admin/appartment', 'index')->name('admin.apartment.index');
-            Route::post('/admin/appartment/create', 'create')->name('appartment.create');
-            Route::get('/admin/appartment/{apartment}/edit', 'edit')->name('appartment.edit');
-            Route::put('/admin/appartment/{apartment}/update', 'update')->name('appartment.update');
-            Route::delete('/admin/appartment/{apartment}/delete', 'delete')->name('appartment.delete');
+            Route::get('/admin/apartment', 'index')->name('admin.apartment.index');
+            Route::post('/admin/apartment/create', 'create')->name('admin.apartment.create');
+            Route::get('/admin/apartment/{apartment}/edit', 'edit')->name('admin.apartment.edit');
+            Route::put('/admin/apartment/{apartment}/update', 'update')->name('admin.apartment.update');
+            Route::delete('/admin/apartment/{apartment}/delete', 'delete')->name('admin.apartment.delete');
         });
         Route::controller(ImageController::class)->group(function() {
             Route::get('/admin/categories/{categoryId}/upload', 'index');
